requirements.txt
streamlit
langchain
langchain-groq
python-dotenv
pydantic
fastapi
uvicorn
python-multipart
markdown
pydantic-settings